Convert module to package
