+++ memory_profiler/__init__.py
@@ -5,20 +5,27 @@ _CLEAN_GLOBALS = globals().copy()
 
 __version__ = '0.55.0'
 
-_CMD_USAGE = "python -m memory_profiler script_file.py"
 
 from functools import wraps
 import inspect
-import linecache
 import logging
 import os
-import pdb
 import subprocess
 import sys
 import time
 import traceback
 import warnings
 import contextlib
+
+from .common import PY2, HAS_TRACEMALLOC
+from .utils import (
+    show_results,
+    choose_backend,
+    get_memory as _get_memory,
+    get_child_memory as _get_child_memory,
+)
+from .line_profiler import LineProfiler
+
 
 if sys.platform == "win32":
     # any value except signal.CTRL_C_EVENT and signal.CTRL_BREAK_EVENT
@@ -27,16 +34,6 @@ if sys.platform == "win32":
 else:
     from signal import SIGKILL
 import psutil
-
-
-if sys.version_info > (3, 3):
-    if sys.version_info < (3, 5):
-        from asyncio import iscoroutinefunction
-    else:
-        from inspect import iscoroutinefunction
-else:
-    def iscoroutinefunction(_):
-        return False
 
 
 # TODO: provide alternative when multiprocessing is not available
@@ -53,10 +50,6 @@ except ImportError:
     line_cell_magic = lambda func: func
     magics_class = lambda cls: cls
 
-PY2 = sys.version_info[0] == 2
-
-_TWO_20 = float(2 ** 20)
-
 if PY2:
     import __builtin__ as builtins
     to_str = lambda x: x
@@ -65,13 +58,8 @@ else:
     import builtins
     to_str = lambda x: str(x)
 
-# .. get available packages ..
-try:
+if HAS_TRACEMALLOC:
     import tracemalloc
-
-    has_tracemalloc = True
-except ImportError:
-    has_tracemalloc = False
 
 
 class MemitResult(object):
@@ -97,108 +85,6 @@ class MemitResult(object):
     def _repr_pretty_(self, p, cycle):
         msg = str(self)
         p.text(u'<MemitResult : ' + msg + u'>')
-
-
-def _get_child_memory(process, meminfo_attr=None):
-    """
-    Returns a generator that yields memory for all child processes.
-    """
-    # Convert a pid to a process
-    if isinstance(process, int):
-        if process == -1: process = os.getpid()
-        process = psutil.Process(process)
-
-    if not meminfo_attr:
-        # Use the psutil 2.0 attr if the older version isn't passed in.
-        meminfo_attr = 'memory_info' if hasattr(process, 'memory_info') else 'get_memory_info'
-
-    # Select the psutil function get the children similar to how we selected
-    # the memory_info attr (a change from excepting the AttributeError).
-    children_attr = 'children' if hasattr(process, 'children') else 'get_children'
-
-    # Loop over the child processes and yield their memory
-    try:
-        for child in getattr(process, children_attr)(recursive=True):
-            yield getattr(child, meminfo_attr)()[0] / _TWO_20
-    except (psutil.NoSuchProcess, psutil.AccessDenied):
-        # https://github.com/fabianp/memory_profiler/issues/71
-        yield 0.0
-
-
-def _get_memory(pid, backend, timestamps=False, include_children=False, filename=None):
-    # .. low function to get memory consumption ..
-    if pid == -1:
-        pid = os.getpid()
-
-    def tracemalloc_tool():
-        # .. cross-platform but but requires Python 3.4 or higher ..
-        stat = next(filter(lambda item: str(item).startswith(filename),
-                           tracemalloc.take_snapshot().statistics('filename')))
-        mem = stat.size / _TWO_20
-        if timestamps:
-            return mem, time.time()
-        else:
-            return mem
-
-    def ps_util_tool():
-        # .. cross-platform but but requires psutil ..
-        process = psutil.Process(pid)
-        try:
-            # avoid using get_memory_info since it does not exists
-            # in psutil > 2.0 and accessing it will cause exception.
-            meminfo_attr = 'memory_info' if hasattr(process, 'memory_info') \
-                else 'get_memory_info'
-            mem = getattr(process, meminfo_attr)()[0] / _TWO_20
-            if include_children:
-                mem +=  sum(_get_child_memory(process, meminfo_attr))
-            if timestamps:
-                return mem, time.time()
-            else:
-                return mem
-        except psutil.AccessDenied:
-            pass
-            # continue and try to get this from ps
-
-    def posix_tool():
-        # .. scary stuff ..
-        if include_children:
-            raise NotImplementedError((
-                "The psutil module is required to monitor the "
-                "memory usage of child processes."
-            ))
-
-        warnings.warn("psutil module not found. memory_profiler will be slow")
-        # ..
-        # .. memory usage in MiB ..
-        # .. this should work on both Mac and Linux ..
-        # .. subprocess.check_output appeared in 2.7, using Popen ..
-        # .. for backwards compatibility ..
-        out = subprocess.Popen(['ps', 'v', '-p', str(pid)],
-                               stdout=subprocess.PIPE
-                               ).communicate()[0].split(b'\n')
-        try:
-            vsz_index = out[0].split().index(b'RSS')
-            mem = float(out[1].split()[vsz_index]) / 1024
-            if timestamps:
-                return mem, time.time()
-            else:
-                return mem
-        except:
-            if timestamps:
-                return -1, time.time()
-            else:
-                return -1
-
-    if backend == 'tracemalloc' and \
-            (filename is None or filename == '<unknown>'):
-        raise RuntimeError(
-            'There is no access to source file of the profiled function'
-        )
-
-    tools = {'tracemalloc': tracemalloc_tool,
-             'psutil': ps_util_tool,
-             'posix': posix_tool}
-    return tools[backend]()
 
 
 class MemTimer(Process):
@@ -458,23 +344,6 @@ def memory_usage(proc=-1, interval=.1, timeout=None, timestamps=False,
 # .. utility functions for line-by-line ..
 
 
-def _find_script(script_name):
-    """ Find the script.
-
-    If the input is not a file, then $PATH will be searched.
-    """
-    if os.path.isfile(script_name):
-        return script_name
-    path = os.getenv('PATH', os.defpath).split(os.pathsep)
-    for folder in path:
-        if not folder:
-            continue
-        fn = os.path.join(folder, script_name)
-        if os.path.isfile(fn):
-            return fn
-
-    sys.stderr.write('Could not find script {0}\n'.format(script_name))
-    raise SystemExit(1)
 
 
 class _TimeStamperCM(object):
@@ -601,234 +470,6 @@ class TimeStamper:
             for ts, level in zip(timestamps, self.stack[func]):
                 stream.write("FUNC %s %.4f %.4f %.4f %.4f %d\n" % (
                     (function_name,) + ts[0] + ts[1] + (level,)))
-
-
-class CodeMap(dict):
-    def __init__(self, include_children, backend):
-        self.include_children = include_children
-        self._toplevel = []
-        self.backend = backend
-
-    def add(self, code, toplevel_code=None):
-        if code in self:
-            return
-
-        if toplevel_code is None:
-            filename = code.co_filename
-            if filename.endswith((".pyc", ".pyo")):
-                filename = filename[:-1]
-            if not os.path.exists(filename):
-                print('ERROR: Could not find file ' + filename)
-                if filename.startswith(("ipython-input", "<ipython-input")):
-                    print(
-                        "NOTE: %mprun can only be used on functions defined in"
-                        " physical files, and not in the IPython environment.")
-                return
-
-            toplevel_code = code
-            (sub_lines, start_line) = inspect.getsourcelines(code)
-            linenos = range(start_line,
-                            start_line + len(sub_lines))
-            self._toplevel.append((filename, code, linenos))
-            self[code] = {}
-        else:
-            self[code] = self[toplevel_code]
-
-        for subcode in filter(inspect.iscode, code.co_consts):
-            self.add(subcode, toplevel_code=toplevel_code)
-
-    def trace(self, code, lineno, prev_lineno):
-        memory = _get_memory(-1, self.backend, include_children=self.include_children,
-                             filename=code.co_filename)
-        prev_value = self[code].get(lineno, None)
-        previous_memory = prev_value[1] if prev_value else 0
-        previous_inc = prev_value[0] if prev_value else 0
-
-        prev_line_value = self[code].get(prev_lineno, None) if prev_lineno else None
-        prev_line_memory = prev_line_value[1] if prev_line_value else 0
-        self[code][lineno] = (max(previous_inc, memory-prev_line_memory), max(memory, previous_memory))
-
-    def items(self):
-        """Iterate on the toplevel code blocks."""
-        for (filename, code, linenos) in self._toplevel:
-            measures = self[code]
-            if not measures:
-                continue  # skip if no measurement
-            line_iterator = ((line, measures.get(line)) for line in linenos)
-            yield (filename, line_iterator)
-
-
-class LineProfiler(object):
-    """ A profiler that records the amount of memory for each line """
-
-    def __init__(self, **kw):
-        include_children = kw.get('include_children', False)
-        backend = kw.get('backend', 'psutil')
-        self.code_map = CodeMap(
-            include_children=include_children, backend=backend)
-        self.enable_count = 0
-        self.max_mem = kw.get('max_mem', None)
-        self.prevlines = []
-        self.backend = choose_backend(kw.get('backend', None))
-        self.prev_lineno = None
-
-    def __call__(self, func=None, precision=1):
-        if func is not None:
-            self.add_function(func)
-            f = self.wrap_function(func)
-            f.__module__ = func.__module__
-            f.__name__ = func.__name__
-            f.__doc__ = func.__doc__
-            f.__dict__.update(getattr(func, '__dict__', {}))
-            return f
-        else:
-            def inner_partial(f):
-                return self.__call__(f, precision=precision)
-
-            return inner_partial
-
-    def add_function(self, func):
-        """ Record line profiling information for the given Python function.
-        """
-        try:
-            # func_code does not exist in Python3
-            code = func.__code__
-        except AttributeError:
-            warnings.warn("Could not extract a code object for the object %r"
-                          % func)
-        else:
-            self.code_map.add(code)
-
-    def wrap_function(self, func):
-        """ Wrap a function to profile it.
-        """
-
-        def f(*args, **kwds):
-            self.enable_by_count()
-            try:
-                return func(*args, **kwds)
-            finally:
-                self.disable_by_count()
-
-        return f
-
-    def runctx(self, cmd, globals, locals):
-        """ Profile a single executable statement in the given namespaces.
-        """
-        self.enable_by_count()
-        try:
-            exec(cmd, globals, locals)
-        finally:
-            self.disable_by_count()
-        return self
-
-    def enable_by_count(self):
-        """ Enable the profiler if it hasn't been enabled before.
-        """
-        if self.enable_count == 0:
-            self.enable()
-        self.enable_count += 1
-
-    def disable_by_count(self):
-        """ Disable the profiler if the number of disable requests matches the
-        number of enable requests.
-        """
-        if self.enable_count > 0:
-            self.enable_count -= 1
-            if self.enable_count == 0:
-                self.disable()
-
-    def trace_memory_usage(self, frame, event, arg):
-        """Callback for sys.settrace"""
-        if frame.f_code in self.code_map:
-            if event == 'call':
-                # "call" event just saves the lineno but not the memory
-                self.prevlines.append(frame.f_lineno)
-            elif event == 'line':
-                # trace needs current line and previous line
-                self.code_map.trace(frame.f_code, self.prevlines[-1], self.prev_lineno)
-                # saving previous line
-                self.prev_lineno = self.prevlines[-1]
-                self.prevlines[-1] = frame.f_lineno
-            elif event == 'return':
-                lineno = self.prevlines.pop()
-                self.code_map.trace(frame.f_code, lineno, self.prev_lineno)
-                self.prev_lineno = lineno
-
-        if self._original_trace_function is not None:
-            self._original_trace_function(frame, event, arg)
-
-        return self.trace_memory_usage
-
-    def trace_max_mem(self, frame, event, arg):
-        # run into PDB as soon as memory is higher than MAX_MEM
-        if event in ('line', 'return') and frame.f_code in self.code_map:
-            c = _get_memory(-1, self.backend, filename=frame.f_code.co_filename)
-            if c >= self.max_mem:
-                t = ('Current memory {0:.2f} MiB exceeded the '
-                     'maximum of {1:.2f} MiB\n'.format(c, self.max_mem))
-                sys.stdout.write(t)
-                sys.stdout.write('Stepping into the debugger \n')
-                frame.f_lineno -= 2
-                p = pdb.Pdb()
-                p.quitting = False
-                p.stopframe = frame
-                p.returnframe = None
-                p.stoplineno = frame.f_lineno - 3
-                p.botframe = None
-                return p.trace_dispatch
-
-        if self._original_trace_function is not None:
-            (self._original_trace_function)(frame, event, arg)
-
-        return self.trace_max_mem
-
-    def __enter__(self):
-        self.enable_by_count()
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        self.disable_by_count()
-
-    def enable(self):
-        self._original_trace_function = sys.gettrace()
-        if self.max_mem is not None:
-            sys.settrace(self.trace_max_mem)
-        else:
-            sys.settrace(self.trace_memory_usage)
-
-    def disable(self):
-        sys.settrace(self._original_trace_function)
-
-
-def show_results(prof, stream=None, precision=1):
-    if stream is None:
-        stream = sys.stdout
-    template = '{0:>6} {1:>12} {2:>12}   {3:<}'
-
-    for (filename, lines) in prof.code_map.items():
-        header = template.format('Line #', 'Mem usage', 'Increment',
-                                 'Line Contents')
-
-        stream.write(u'Filename: ' + filename + '\n\n')
-        stream.write(header + u'\n')
-        stream.write(u'=' * len(header) + '\n')
-
-        all_lines = linecache.getlines(filename)
-
-        float_format = u'{0}.{1}f'.format(precision + 4, precision)
-        template_mem = u'{0:' + float_format + '} MiB'
-        for (lineno, mem) in lines:
-            if mem:
-                inc = mem[0]
-                mem = mem[1]
-                mem = template_mem.format(mem)
-                inc = template_mem.format(inc)
-            else:
-                mem = u''
-                inc = u''
-            tmp = template.format(lineno, mem, inc, all_lines[lineno - 1])
-            stream.write(to_str(tmp))
-        stream.write(u'\n\n')
 
 
 def _func_exec(stmt, ns):
@@ -1113,7 +754,7 @@ def profile(func=None, stream=None, precision=1, backend='psutil'):
     Decorator that will run the function and print a line-by-line profile
     """
     backend = choose_backend(backend)
-    if backend == 'tracemalloc' and has_tracemalloc:
+    if backend == 'tracemalloc' and HAS_TRACEMALLOC:
         if not tracemalloc.is_tracing():
             tracemalloc.start()
     if func is not None:
@@ -1133,58 +774,32 @@ def profile(func=None, stream=None, precision=1, backend='psutil'):
         return inner_wrapper
 
 
-def choose_backend(new_backend=None):
-    """
-    Function that tries to setup backend, chosen by user, and if failed,
-    setup one of the allowable backends
-    """
-
-    _backend = 'no_backend'
-    all_backends = [
-        ('psutil', True),
-        ('posix', os.name == 'posix'),
-        ('tracemalloc', has_tracemalloc),
-    ]
-    backends_indices = dict((b[0], i) for i, b in enumerate(all_backends))
-
-    if new_backend is not None:
-        all_backends.insert(0, all_backends.pop(backends_indices[new_backend]))
-
-    for n_backend, is_available in all_backends:
-        if is_available:
-            _backend = n_backend
-            break
-    if _backend != new_backend and new_backend is not None:
-        warnings.warn('{0} can not be used, {1} used instead'.format(
-            new_backend, _backend))
-    return _backend
-
-
 # Insert in the built-ins to have profile
 # globally defined (global variables is not enough
 # for all cases, e.g. a script that imports another
 # script where @profile is used)
 def exec_with_profiler(filename, profiler, backend, passed_args=[]):
-    from runpy import run_module
     builtins.__dict__['profile'] = profiler
-    ns = dict(_CLEAN_GLOBALS,
-              profile=profiler,
-             # Make sure the __file__ variable is usable
-             # by the script we're profiling
-              __file__=filename)
+    ns = dict(
+        _CLEAN_GLOBALS,
+        profile=profiler,
+        # Make sure the __file__ variable is usable
+        # by the script we're profiling
+        __file__=filename
+    )
     # Make sure the script's directory in on sys.path
     # credit to line_profiler
-    sys.path.insert(0, os.path.dirname(script_filename))
+    sys.path.insert(0, os.path.dirname(filename))
 
     _backend = choose_backend(backend)
     sys.argv = [filename] + passed_args
     try:
-        if _backend == 'tracemalloc' and has_tracemalloc:
+        if _backend == 'tracemalloc' and HAS_TRACEMALLOC:
             tracemalloc.start()
         with open(filename) as f:
             exec(compile(f.read(), filename, 'exec'), ns, ns)
     finally:
-        if has_tracemalloc and tracemalloc.is_tracing():
+        if HAS_TRACEMALLOC and tracemalloc.is_tracing():
             tracemalloc.stop()
 
 
@@ -1197,12 +812,12 @@ def run_module_with_profiler(module, profiler, backend, passed_args=[]):
     if PY2:
         run_module(module, run_name="__main__", init_globals=ns)
     else:
-        if _backend == 'tracemalloc' and has_tracemalloc:
+        if _backend == 'tracemalloc' and HAS_TRACEMALLOC:
             tracemalloc.start()
         try:
             run_module(module, run_name="__main__", init_globals=ns)
         finally:
-            if has_tracemalloc and tracemalloc.is_tracing():
+            if HAS_TRACEMALLOC and tracemalloc.is_tracing():
                 tracemalloc.stop()
 
 
@@ -1236,61 +851,3 @@ class LogFile(object):
     def flush(self):
         for handler in self.logger.handlers:
             handler.flush()
-
-
-if __name__ == '__main__':
-    from argparse import ArgumentParser, REMAINDER
-
-    parser = ArgumentParser(usage=_CMD_USAGE)
-    parser.add_argument('--version', action='version', version=__version__)
-    parser.add_argument(
-        '--pdb-mmem', dest='max_mem', metavar='MAXMEM',
-        type=float, action='store',
-        help='step into the debugger when memory exceeds MAXMEM')
-    parser.add_argument(
-        '--precision', dest='precision', type=int,
-        action='store', default=3,
-        help='precision of memory output in number of significant digits')
-    parser.add_argument('-o', dest='out_filename', type=str,
-        action='store', default=None,
-        help='path to a file where results will be written')
-    parser.add_argument('--timestamp', dest='timestamp', default=False,
-        action='store_true',
-        help='''print timestamp instead of memory measurement for
-        decorated functions''')
-    parser.add_argument('--backend', dest='backend', type=str, action='store',
-        choices=['tracemalloc', 'psutil', 'posix'], default='psutil',
-        help='backend using for getting memory info '
-             '(one of the {tracemalloc, psutil, posix})')
-    parser.add_argument("program", nargs=REMAINDER,
-        help='python script or module followed by command line arguements to run')
-    args = parser.parse_args()
-
-    if len(args.program) == 0:
-        print("A program to run must be provided. Use -h for help")
-        sys.exit(1)
-
-    target = args.program[0]
-    script_args = args.program[1:]
-    _backend = choose_backend(args.backend)
-    if args.timestamp:
-        prof = TimeStamper(_backend)
-    else:
-        prof = LineProfiler(max_mem=args.max_mem, backend=_backend)
-
-    try:
-        if args.program[0].endswith('.py'):
-            script_filename = _find_script(args.program[0])
-            exec_with_profiler(script_filename, prof, args.backend, script_args)
-        else:
-            run_module_with_profiler(target, prof, args.backend, script_args)
-    finally:
-        if args.out_filename is not None:
-            out_file = open(args.out_filename, "a")
-        else:
-            out_file = sys.stdout
-
-        if args.timestamp:
-            prof.show_results(stream=out_file)
-        else:
-            show_results(prof, precision=args.precision, stream=out_file)
